+++ src/lichess_board/__init__.py
@@ -1,4 +1,0 @@
-from .__version__ import __version__
-from .widget import ChessBoardWidget
-
-__all__ = ["ChessBoardWidget", "__version__"]